requirements.txt
pyzbar
pillow
